feat: add GGUF model support to CLI and server
Both tools now accept a .gguf file path in --model-dir. Detection is by
.gguf extension; checkpoint directories work exactly as before.

Changes:
- BPETokenizer::from_vocab(): new factory to build tokenizer from a GGUF
  file's embedded vocab (id→string list) and merge pairs, used by both
  tools when loading GGUF models
- tools/cli: branch on is_gguf_path(); load via GGUFReader + load_gguf_model()
  and construct tokenizer from reader.vocab()
- tools/server: same in InferenceEngine constructor; step_ set to 0 for
  GGUF (no training step in file); model display shows file path
- Fixed pre-existing -Wsign-conversion warnings surfaced by bpe.hpp recompile:
  ch06_embeddings (span[] indexing with int64_t), ch24_real_training
  (int64_t * int64_t → double), bpe.cpp (unsigned char cast, resize cast)

Usage:
  sub0llm-cli   --model-dir model.gguf [opts]
  sub0llm-server --model-dir model.gguf [opts]

+++ chapters/ch24_real_training/main.cpp
@@ -586,7 +586,7 @@ static void section_full_training(int steps, const std::string& ckpt_dir, bool d
             const double tok_s  = elapsed > 0 ? static_cast<double>(tokens_trained) / elapsed : 0.0;
             const int64_t remaining = static_cast<int64_t>(steps) - step - 1;
             const double eta_sec = (tok_s > 0 && remaining > 0)
-                ? static_cast<double>(remaining) * seq_len / tok_s : 0.0;
+                ? static_cast<double>(remaining) * static_cast<double>(seq_len) / tok_s : 0.0;
             std::string eta_str = eta_sec > 3600 ? std::format("{:.1f}h", eta_sec / 3600)
                                 : eta_sec > 60   ? std::format("{:.1f}m", eta_sec / 60)
                                 :                  std::format("{:.0f}s", eta_sec);
